--- packages/html-to-slate-ast/src/client/index.ts
@@ -1,24 +1,0 @@
-import { Descendant } from 'slate';
-import { normalizeHtml } from '../normalizeHtml';
-import { deserialize } from '../deserialize';
-
-const parseDomDocumentSync = (normalizedHTML: string) => {
-  return new DOMParser().parseFromString(normalizedHTML, 'text/html');
-};
-
-export function htmlToSlateASTSync<T>(
-  html: string
-): string | Descendant | ChildNode[] | Descendant[] | T | T[] {
-  if (
-    typeof window === 'undefined' ||
-    typeof window.DOMParser === 'undefined'
-  ) {
-    throw new Error(
-      'This function is intended to be used in a browser environment only'
-    );
-  }
-
-  const normalizedHTML = normalizeHtml(html);
-  const domDocument = parseDomDocumentSync(normalizedHTML);
-  return deserialize(domDocument.body, window);
-}
